requirements.txt
pandas
openai
python-dotenv
huggingface_hub
pyarrow
maritaca-python
maritalk
groq
googletrans
tqdm